requirements.txt
streamlit
opencv-python-headless
ultralytics
pyserial
firebase-admin
pandas
plotly
torch
torchvision
torchaudio